day4_sam2/requirements.txt
ultralytics==8.2.28
gradio==4.36.1
gradio-client==0.17.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0